Final report Jupter notebook
Here we looked at the UCI heart dataset which is used to predict heart disease with machine learning. Python was the language of choice here and we used scikit-learn for machine learning models, statsmodel for computing basic statistics, matplotlib and seaborn for data visualization and numpy for basic array work.